Imporve perf of random read and insert comapre by suggesting   function inlining.
